Fixed the refactor problem and now the mod is `apollomod`! :D
