+++ src/main/resources/assets/apollomod/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "item.apollomod.copper": "Copper",
+  "item.apollomod.ruby": "Ruby"
+}
