+++ src/main/resources/assets/apollomod/models/item/copper.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "examplemod:items/ruby"
+    "layer0": "apollomod:items/copper"
   }
 }
