+++ src/main/resources/assets/apollomod/models/item/ruby.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "examplemod:items/copper"
+    "layer0": "apollomod:items/ruby"
   }
 }
