+++ src/main/resources/assets/apollomod/textures/items/copper.png
