+++ src/main/resources/assets/apollomod/textures/items/ruby.png
