--- src/main/resources/assets/examplemod/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "item.examplemod.copper": "Copper",
-  "item.examplemod.ruby": "Ruby"
-}
